Refactor panels
